Exclude live_preview from 403 handling

+++ src/common/helpers/request.ts
@@ -20,9 +20,19 @@ client.interceptors.response.use(
     return response;
   },
   (error: AxiosError<ValidationBag>) => {
+    console.log(error.request.responseURL);
+    console.log(error.response);
+
     if (error.response?.status === 403) {
-      window.location.reload();
-      localStorage.clear();
+      const url = new URL(error.request.responseURL);
+
+      // Temporary solution. If you see this in few months,
+      // please delete.
+
+      if (!url.pathname.includes('/api/v1/live_preview')) {
+        window.location.reload();
+        localStorage.clear();
+      }
     }
 
     if (error.response?.status === 404) {
